ImpactX 25.10: Fix CSV
For once in a long time, we will do a small breaking change
in ImpactX, in particular in its CSV files for reduced beam
diagnostics: we unify the column labels.

Once ImpactX 25.10 will be released in October, this patch
will make sure ABEL will use the new column names automatically.